Separate mask retreval for MPASO and MPASSI
The MPAS-Seaice restart files don't include the maxLevelCell
variable that is used in MPAS-Ocean to determine the 2D cell
mask, and certainly don't include the information needed to
determine the 3D mask (which only applies to the ocean).  As a
result, it makes sense to have different masks for the 2
components.